:boom: Page to view public recipes
Recipe model was modified so now they can be public to other users. So, we created a endpoint '/api/recipes' that will display all public recipes from all users. The client will use this endpoint to show to users recipes from others users (if they are public). Add/Edit page were edited to support this new feature. The next feature will be to show the name of the chef that created the recipe if the recipe is public.